Allow to install captain by major version lookup (#5)
This change allows a user to install the latest `v1.x.x` version of the
CLI by referring to it as `v1`. Also, this is now the new default when
no version number was supplied. Explicitly setting `latest` as the
version number continues to work.

> **Warning**: This is currently broken and should not be merged until
we actually have a `v1` version of the CLI released.

Further, this technically constitutes a breaking change to the public
API of this action and would require a major version change as we follow
semantic versioning. However, the main motivation behind changing the
default behaviour was to align the `v1` version of this action with the
`v1` release of the CLI, so a major version bump would defeat the
purpose here.

Co-authored-by: Dan Manges <dan@rwx.com>